Improvements in yatester
* Moved /throw to selftest
* Parser buffers are now resizable
* Command argument count is now unsigned (size_t)
* Removed argtest and moved its tests to selftest

+++ src/yatester/cmd.h
@@ -3,10 +3,12 @@
 
 /* Commands */
 
+#include <stddef.h>
+
 typedef struct
 {
 	const char *name;
-	int argc;
+	size_t argc;
 	void (*handler)(const char** argv);
 }
 yatester_command;
