+++ src/yatester/runner.h
@@ -5,6 +5,8 @@
 
 #include <yatester/err.h>
 
-yatester_status yatester_runcommand(const char* commandname, int argc, const char** argv);
+#include <stddef.h>
+
+yatester_status yatester_runcommand(const char* commandname, size_t argc, const char** argv);
 
 #endif
